requirements.txt
numpy
pandas
scikit-learn
pickle
joblib
pillow
nltk